+++ docs/src/components/CodeHighlight.tsx
@@ -43,17 +43,9 @@ export const CodeHighlight = ({
           >
             <View as="code" className={classNames(className, prismClassName)}>
               {tokens.map((line, i) => (
-                <View
-                  className="code-line"
-                  key={i}
-                  {...getLineProps({ line, key: i })}
-                >
+                <View {...getLineProps({ line })} className="code-line" key={i}>
                   {line.map((token, key) => (
-                    <Text
-                      as="span"
-                      key={key}
-                      {...getTokenProps({ token, key })}
-                    />
+                    <Text {...getTokenProps({ token })} as="span" key={key} />
                   ))}
                 </View>
               ))}