+++ packages/react-notifications/src/components/InAppMessaging/Backdrop/Backdrop.tsx
@@ -6,7 +6,10 @@ import { BackdropProps } from './types';
 
 export const BACKDROP_TEST_ID = 'inappmessaging-backdrop';
 
-export function Backdrop({ onClick, ...rest }: BackdropProps): JSX.Element {
+export function Backdrop({
+  onClick,
+  ...rest
+}: BackdropProps): React.JSX.Element {
   return (
     <View
       className="amplify-inappmessaging-backdrop"