+++ packages/react-native-auth/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @aws-amplify/ui-react-native-auth
 
+## 0.0.42
+
+### Patch Changes
+
+- Updated dependencies [[`272ec8e575718ea990dcb9f51d1baaf522f45fb6`](https://github.com/aws-amplify/amplify-ui/commit/272ec8e575718ea990dcb9f51d1baaf522f45fb6)]:
+  - @aws-amplify/ui@6.8.1
+  - @aws-amplify/ui-react-core-auth@0.0.39
+  - @aws-amplify/ui-react-native@2.4.1
+
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [[`8757f03b7b1d761d9a13cfe29382a2208d4f3999`](https://github.com/aws-amplify/amplify-ui/commit/8757f03b7b1d761d9a13cfe29382a2208d4f3999), [`8d2aa79b174d4e48aaca913bd568898b8c8975a2`](https://github.com/aws-amplify/amplify-ui/commit/8d2aa79b174d4e48aaca913bd568898b8c8975a2)]:
+  - @aws-amplify/ui@6.8.0
+  - @aws-amplify/ui-react-native@2.4.0
+  - @aws-amplify/ui-react-core-auth@0.0.38
+
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [[`475e4a37a404bd116ec55e95579acb8f2d23c9a4`](https://github.com/aws-amplify/amplify-ui/commit/475e4a37a404bd116ec55e95579acb8f2d23c9a4)]:
+  - @aws-amplify/ui-react-native@2.3.1
+  - @aws-amplify/ui-react-core-auth@0.0.37
+
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [[`825cccc92f5c326f0f110c8439a1c2c6553ff00a`](https://github.com/aws-amplify/amplify-ui/commit/825cccc92f5c326f0f110c8439a1c2c6553ff00a), [`50cde53145bfffc97f17b6da5c650c25d992fd97`](https://github.com/aws-amplify/amplify-ui/commit/50cde53145bfffc97f17b6da5c650c25d992fd97), [`954e9be2d055a57ebddab5b902e3e6b427d09f67`](https://github.com/aws-amplify/amplify-ui/commit/954e9be2d055a57ebddab5b902e3e6b427d09f67)]:
+  - @aws-amplify/ui-react-native@2.3.0
+  - @aws-amplify/ui-react-core-auth@0.0.36
+
 ## 0.0.38
 
 ### Patch Changes